slight changes to the TPS Camera
Sorry for messing with it, but now the camera wont clip through walls. The reticule needs adjustment. On play it snaps to the players head, by looking at the scene view and the game view, it looks like its not matching up with the gun. shooting still follows the reticule, I'm not sure what it looked like before. But we can work together to sort it out if need be.