EnemyAI jump and dash attacks
Code is working for the jumpattack and dashattack. Though perhaps I was a little too high with my Range numbers as there are some kinks needed to be worked out.

For example, after setting up the different enemy types in the inspector for these 2 attacks, the mythril geistwalker, jumps way higher than it should and when landing, disappears through the ground.

Also need to add audio and animations.