Dynamic Music Start, Store Rework, UI Changes And Adjustments
Dynamic Music is working, it just needs to be implemented fully across scenes and linked with appropriate music, I still need to adjust the Battle Music and cut off the first 49 seconds, easy. I'm reworking the store in an attempt to ensure it is ready, with Kylars permission, it is an undertaking, I'm also adding save/load to everything store related. I prefabs the singletons and placed them on the Main Menu Scene, they persist and there only needs to be one of each, so I removed the rest to clean it up and remove errors in the console. Will keep working.