Persistence Fix, Volume Slider Fix, User Settings Fix
Fixed persistence issues by forcing all persistence to be destroyed on new game and restart. Fixed volume slider and persistence of user settings as far as I can tell. Needs more testing to be sure, but it seems to be working in my workspace and I've applied it to the game. Music still needs to be adjusted per scene or menu, but the work is done, it should just be tweaking things now.